Remove Mod_Meta_L_Disable
It seems this was a feature of keyboard.pm (look at the drakx svn log and
ChangeLog), but I just can't find a way to set it with the current code (and
yes, I grep'ed a lot of code).  If we ever re-add this script, we should add it
to the package that uses it, not to xinitrc (or even consider implementing the
script somehow else, since it runs every time every Mandriva user logs in).


SILENT: svn-revision: 638603